Add support for Elliptic Curve Diffie-Hellman.
This patch adds support for elliptic curve Diffie-Hellman for ephemeral
key exchange between client and server.

We add the configuration option `ecdh_curve` in the `serverinfo` block
to allow the user to specify a preferred curve.

Thanks to Yawning on #tor-dev at OFTC for answering the questions I have
had.