Apply caller ID logic to invites as well (#52)
* Apply caller ID logic to invites as well

This change extends the  existing caller ID logic that is applied when
either +g or +G to apply to invites as well. It can be bypassed by opers
if the opers_bypass_callerid config option is set.

* Send invite notification when caller ID is enabled

Send a "is messaging or inviting you" message to the target when caller
ID is enabled and an invite is blocked. It uses the same caller ID rate
limiting as messages.